Decrement line number in reconsume().
The oversight in the line number pull request was when having an unescaped
newline in either a single-quoted or a doubly-quoted string, such as:

    foo {
      bar: "foo
    }

In that example, the closing curly brace was on line 3 instead of 2
(line numbers are 0-indexed).

In addition, I use this modification to eradicate overlapping line information
that occurred in neighbouring tokens.